GG-25278: fix platform #2

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/platform/cache/PlatformCache.java
@@ -1450,12 +1450,12 @@ public class PlatformCache extends PlatformAbstractTarget {
         boolean distrJoins = reader.readBoolean();
         boolean enforceJoinOrder = reader.readBoolean();
         boolean lazy = reader.readBoolean();
-        long timeout = reader.readLong();
+        int timeout = reader.readInt();
         boolean replicated = reader.readBoolean();
         boolean collocated = reader.readBoolean();
         String schema = reader.readString();
 
-        SqlFieldsQuery qry = QueryUtils.withQueryTimeout(new SqlFieldsQuery(sql), (int)timeout, TimeUnit.SECONDS)
+        SqlFieldsQuery qry = QueryUtils.withQueryTimeout(new SqlFieldsQuery(sql), timeout, TimeUnit.SECONDS)
             .setPageSize(pageSize)
             .setArgs(args)
             .setLocal(loc)
